Bugfix: graphicspath ignored in IfFileExists
It's a known issue that IfFileExists ignores graphicspath, since the
latter is only updated once an image is included. We ran into this
problem when adding logos and signatures where not placed in the TDS but
added via graphicspath (since we check if logos and signatures exist
before trying to include them via IfFileExists). The solution is due to
Daniel Diniz on stackexchange:

https://tex.stackexchange.com/questions/341526/how-to-use-graphicspath-in-iffileexists/

Closes: #12