Change LICENSE to contain CC0 text, update README
We've decided to move to "unlicense"/CC0 the classes and the new
copyright notices are based on the suggestions by the Creative Commons
organization for how to do that.

For details, see
https://creativecommons.org/share-your-work/public-domain/cc0/

and, in particular, the suggestions concerning software in the FAQ:

https://wiki.creativecommons.org/wiki/CC0_FAQ#May_I_apply_CC0_to_computer_software.3F_If_so.2C_is_there_a_recommended_implementation.3F

Even though CC0 is not a license, we've included the text in
```LICENSE``` because GitHub reads this file. This is why we decided
against the advice of the CC to call the file (more appropriately)
```COPYING```.

Co-authored-by: Keven Quach <belfami@gmail.com>